notifications-bot/requirements.txt
aiogram>=3.0.0
redis>=5.0.0
python-dotenv 